Properly fix react-refresh warnings by separating concerns
Instead of disabling ESLint rules, properly fixed the architectural issues:

- Moved context definitions to separate files in src/contexts/
- Moved hooks to separate files in src/hooks/
- Moved constants to src/constants/solidProviders.ts
- Component files now only export components (React Fast Refresh compatible)
- Hooks are in their own files and can be imported independently
- All imports updated across the codebase

This follows React Fast Refresh best practices where files should export
either components OR hooks/constants, not a mix of both.
