+++ src/pages/edit-questions-form.tsx
@@ -7,12 +7,12 @@ import { DatabaseMigration } from '../services/migration'
 import { QuestionSection } from '../edit-questions/question-section'
 import { PeopleSection } from '../edit-questions/people-section'
 import { Button } from '../components/Button'
-import { useToast } from '../components/ToastContext'
+import { useToast } from '../hooks/useToast'
 import { AlwaysNeededItemsSection } from '../edit-questions/always-needed-items-section'
 import { Modal } from '../components/Modal'
 import { exampleData } from '../edit-questions/example-data'
 import { Callout } from '../components/Callout'
-import { useSolidPod } from '../components/SolidPodContext'
+import { useSolidPod } from '../hooks/useSolidPod'
 import { usePodSync } from '../hooks/usePodSync'
 import { useSyncCoordinator } from '../hooks/useSyncCoordinator'
 import { POD_CONTAINERS } from '../services/solidPod'
