+++ src/pages/useWizardGeneration.ts
@@ -1,5 +1,5 @@
 import { useState } from 'react'
-import { useToast } from '../components/ToastContext'
+import { useToast } from '../hooks/useToast'
 import { packingAppDb } from '../services/database'
 import { createExampleData } from '../edit-questions/example-data'
 import { QUESTION_SET_ID } from '../constants'
