+++ src/pages/view-packing-list.tsx
@@ -5,8 +5,8 @@ import { PackingList } from '../create-packing-list/types'
 import { packingAppDb } from '../services/database'
 import { Button } from '../components/Button'
 import { useForm, useWatch } from 'react-hook-form'
-import { useSolidPod } from '../components/SolidPodContext'
-import { useToast } from '../components/ToastContext'
+import { useSolidPod } from '../hooks/useSolidPod'
+import { useToast } from '../hooks/useToast'
 import { usePodSync } from '../hooks/usePodSync'
 import { useSyncCoordinator } from '../hooks/useSyncCoordinator'
 import { POD_CONTAINERS } from '../services/solidPod'
